+++ src/main.cpp
@@ -52,17 +52,14 @@ class $modify(TheTransformCtrls, GJTransformControl) {
 	};
 
 	void updateValidSprites() {
-		log::debug("updatevalidsprites called");
 		if (m_fields->warpSprites->count() > 0) {
 			m_fields->warpSprites->removeAllObjects(); // Clear existing objects safely
 		}
-		log::debug("removed objects");
 
 		const std::string textureNomenclature = "warpBtn_02_001.png";
 
 		auto* cachedFrames = CCSpriteFrameCache::sharedSpriteFrameCache()->m_pSpriteFrames;
 		std::unordered_map<std::string, CCSpriteFrame*> frameMap;
-		log::debug("did some things");
 
 		// Populate the frame map with frames matching the nomenclature
 		for (const auto& pair : CCDictionaryExt<std::string, CCSpriteFrame*>(cachedFrames)) {
@@ -73,7 +70,6 @@ class $modify(TheTransformCtrls, GJTransformControl) {
 				frameMap[key] = frame;
 			}
 		}
-		log::debug("populated frame map");
 
 		// Iterate over the main node's children and filter based on texture
 		CCObject* obj;
@@ -87,7 +83,6 @@ class $modify(TheTransformCtrls, GJTransformControl) {
 				}
 			}
 		}
-		log::debug("iterated over main node's children");
 
 		if (m_fields->warpSprites->count() <= 0) {
 			log::warn("updateValidSprites(): No warp sprite textures found. This shouldn't happen, report if this does.");
@@ -98,79 +93,57 @@ class $modify(TheTransformCtrls, GJTransformControl) {
 
 
 	void enableWarpers() {
-		log::debug("enableWarpers(): called.");
 		if (m_fields->disabledWarps->count() > 0) {
-			log::debug("enableWarpers(): more than 0 disabled warpers.");
 			//Disable the disabled. I hope that makes sense :D
 			CCObject* warper;
-			log::debug("enableWarpers(): gettin ready for the ccarray foreach");
 			CCARRAY_FOREACH(m_fields->disabledWarps, warper) {
-				log::debug("enableWarpers(): checking if {} is a valid sprite", warper);
 				if (auto warperSprite = dynamic_cast<CCSprite*>(warper)) {
-					log::debug("enableWarpers(): seems like {} is a valid sprite", warper);
 					warperSprite->setColor({ 255, 255, 255 });
-					log::debug("set color of the aformentioned warper");
 				}
 				else {
 					log::warn("enableWarpers(): Unexpected object inside disabledWarps. This shouldn't happen, please report!!");
 				}
 			}
-			log::debug("enableWarpers(): removing the disabled warps from the array");
 			m_fields->disabledWarps->removeAllObjects();
 		}
 	}
 
 
 	void updateDisabledWarps() {
-		log::debug("updatedDisabledWarps(): got called, enabling warpers.");
 		enableWarpers();
-		log::debug("updatedDisabledWarps(): enabled the warpers");
+
 		if (m_fields->snappedTo == nullptr) {
-			log::debug("updatedDisabledWarps(): not snapped to anything. aborting.");
 			return;
 		}
 
-		log::debug("updatedDisabledWarps(): defining disabledclr");
 		ccColor3B disabledclr = { 140, 90, 90 };
-		log::debug("updatedDisabledWarps(): defining the warpsprite we're snapped to from the mfields");
 		CCSprite* warpSprite = m_fields->snappedTo;
 
-		log::debug("updatedDisabledWarps(): getting the positions");
 		float yPos = warpSprite->getPositionY();
 		float xPos = warpSprite->getPositionX();
 
-		log::debug("updatedDisabledWarps(): defining the amount of whatever i dont wanna type this out");
 		int xCount = 0;
 		int yCount = 0;
 
-		log::debug("updatedDisabledWarps(): gettin ready for another ccarray foreach!");
 		CCObject* warpSpriteObj;
 		CCArray* axisAlignedSprites = CCArray::create(); // WILL contain all sprites that align on either the x or y axis
-		log::debug("updatedDisabledWarps(): Were getting into the danger zone. Created a CCArray");
 		axisAlignedSprites->retain();
-		log::debug("updatedDisabledWarps(): Retained the new ccarray. This might be the crash. Please be. Fgs.");
 		CCARRAY_FOREACH(m_fields->warpSprites, warpSpriteObj) {
-			log::debug("updatedDisabledWarps(): initiated the ccarray. object being checked: {}", warpSpriteObj);
 			if (CCSprite* CCwarpSprite = dynamic_cast<CCSprite*>(warpSpriteObj)) {
-				log::debug("updatedDisabledWarps(): the aformentioned object has passed the dynamiccast.");
 				
 				if (!m_fields->disabledWarps->containsObject(CCwarpSprite) && CCwarpSprite != warpSprite) {
-					log::debug("updatedDisabledWarps(): the aformentioned object has passed another test!");
 					if (abs(CCwarpSprite->getPositionX() - xPos) < EPSILON) {
-						log::debug("updatedDisabledWarps(): the aformentioned object is aligned with something on x idfk");
+
 						axisAlignedSprites->addObject(CCwarpSprite);
-						log::debug("updatedDisabledWarps(): added that object to some array");
 						xCount++;
-						log::debug("updatedDisabledWarps(): incremented x");
 
 					}
 
 					if (abs(CCwarpSprite->getPositionY() - yPos) < EPSILON) {
-						log::debug("updatedDisabledWarps(): the aformentioned object is aligned with something on x idfk");
+
 						axisAlignedSprites->addObject(CCwarpSprite);
-						log::debug("updatedDisabledWarps(): added that object to some array");
 						yCount++;
-						log::debug("updatedDisabledWarps(): incremented y");
+
 					}
 
 				};
@@ -179,21 +152,16 @@ class $modify(TheTransformCtrls, GJTransformControl) {
 		}
 
 		if (axisAlignedSprites->count() != 4) { //If it's not a corner
-			log::debug("updatedDisabledWarps(): its not a corner!");
+
 			if (yCount > xCount) { //if there were more horizontal positions found than vertical ones, it must be a horizontal row
-				log::debug("updatedDisabledWarps(): horiz");
 
 				m_fields->disabledWarps->removeAllObjects();
-				log::debug("updatedDisabledWarps(): horiz, killed em objects, and am gettin ready for yet another ccarray foreachh");
 				CCObject* rowObject;
 				CCARRAY_FOREACH(m_fields->warpSprites, rowObject) {
-					log::debug("updatedDisabledWarps(): horiz, jiasdgjisd");
+
 					if (CCSprite* rowObj = dynamic_cast<CCSprite*>(rowObject)) {
-						log::debug("updatedDisabledWarps(): horiz, i dont understand my own code anymore, {}", rowObj);
 						if (abs(rowObj->getPositionY() - yPos) < EPSILON && rowObj != warpSprite) {
-							log::debug("updatedDisabledWarps(): horiz, yea something matches up, adding it to some array");
 							m_fields->disabledWarps->addObject(rowObj);
-							log::debug("updatedDisabledWarps(): horiz, i think i did that im gonna go insane");
 						}
 					}
 
@@ -201,20 +169,14 @@ class $modify(TheTransformCtrls, GJTransformControl) {
 
 			}
 			else if (xCount > yCount) { // if there were more vertical positions found, must be a vertical row then, right?
-				log::debug("updatedDisabledWarps(): vertical");
 
 				m_fields->disabledWarps->removeAllObjects();
-				log::debug("updatedDisabledWarps(): vertic, killed dem objects, and am gettin ready for yet another ccarray foreachh");
 				CCObject* columnObject;
 				CCARRAY_FOREACH(m_fields->warpSprites, columnObject) {
-					log::debug("updatedDisabledWarps(): vertic, obj is in the ccarray, {}", columnObject);
 
 					if (CCSprite* columnObj = dynamic_cast<CCSprite*>(columnObject)) {
-						log::debug("updatedDisabledWarps(): vertic, columbnobj passed some dynacmiccast");
 						if (abs(columnObj->getPositionX() - xPos) < EPSILON && columnObj != warpSprite) {
-							log::debug("updatedDisabledWarps(): vertic, PASSED ABS....");
 							m_fields->disabledWarps->addObject(columnObj);
-							log::debug("updatedDisabledWarps(): vertic, added object to clumnom...");
 						}
 					}
 
@@ -224,14 +186,10 @@ class $modify(TheTransformCtrls, GJTransformControl) {
 
 		}
 		else {
-			log::debug("hi vitox");
 			CCObject* axisObj;
 			CCARRAY_FOREACH(axisAlignedSprites, axisObj) {
-				log::debug("thanks for helping me with fixing this");
 				if (CCSprite* axisObject = dynamic_cast<CCSprite*>(axisObj)) {
-					log::debug("i appreciate you taking your time with this");
 					m_fields->disabledWarps->addObject(axisObject);
-					log::debug("especially cuz i keep on sending you random files");
 				}
 			}
 
@@ -239,24 +197,19 @@ class $modify(TheTransformCtrls, GJTransformControl) {
 
 		// applying colors because visual stuff gives me dopamine
 		CCObject* warp;
-		log::debug("Look at those disabled warps: {}", m_fields->disabledWarps);
 		CCARRAY_FOREACH(m_fields->disabledWarps, warp) {
-			log::debug("dude im ccarraying something aggain, look, {}", warp);
+
 			if (CCSprite* warpObject = dynamic_cast<CCSprite*>(warp)) {
-				log::debug("and that object seems to actually be a ccsprite");
 				warpObject->setColor(disabledclr);
-				log::debug("im sorry for your console being this clogged.");
 			}
 			
 		}
 		// It's as shrimple as that!!!
 
-		log::debug("releaseing memory. im finally done with typing. u agthe tit");
 		axisAlignedSprites->release(); // memory not leaking anymor
 	};
 
 	std::pair<bool, CCSprite*> snap(bool test) {
-		log::debug("snap called!");
 
 		if (test) { // Testing mode: Returns if the pivot WOULD snap or not. Im repeating lots of code but i dont wanna make a separate function for this
 			if (!m_fields->initialized || !m_mainNodeParent->isVisible()) {
@@ -296,60 +249,52 @@ class $modify(TheTransformCtrls, GJTransformControl) {
 
 		}
 		else {
-			log::debug("updating valid sprites");
+
 			updateValidSprites();
-			log::debug("snap(): updated the valid sprites.");
+
 
 			if (!m_mainNodeParent->isVisible()) {
 				return std::make_pair(false, nullptr);
-				log::debug("mainnodeparent is invisible");
+
 			}
 
 
 			CCSprite* pivotNode = GJTransformControl::spriteByTag(1);
-			log::debug("gjt spritebytag1");
 			CCArray* targets = m_fields->warpSprites;
 			CCRect pivotBox = pivotNode->boundingBox();
 
 			int foundObjs = 0;
 
-			log::debug("gettin ready for the ccarray foreach");
+
 			CCObject* obj;
 			CCARRAY_FOREACH(targets, obj) {
-				log::debug("im looking if {} is touching the pivot node", obj);
 
 				if (CCSprite* warpSprite = dynamic_cast<CCSprite*>(obj)) {
-					log::debug("correctly dynamically casted.");
+
 					if (warpSprite && warpSprite != pivotNode && pivotBox.intersectsRect(warpSprite->boundingBox())) {
-						log::debug("it passed various tests just now");
+
 						CCPoint result = warpSprite->getParent()->convertToWorldSpace(warpSprite->getPosition());
-						log::debug("i got the result position");
 
 						pivotNode->setPosition(pivotNode->getParent()->convertToNodeSpace(result));
-						log::debug("i set position of pivotnode");
 						m_fields->snappedTo = warpSprite;
-						log::debug("set internal value for where the thing is snapped to. magic snapple");
 						foundObjs++;
-						log::debug("incremented found objects");
 						break;
-						log::debug("cancelling, cuz i found snap object");
+
 					}
 				}
 
 			}
-			log::debug("something inbetween! doing some if statement now.");
+
 			if (foundObjs == 0) {
-				log::debug("there are no found objects");
+
 				m_fields->snappedTo = nullptr;
-				log::debug("reset the snappedTo internal value. calling enableWarpers()");
+
 				enableWarpers();
-				log::debug("enableWarpers should be done.");
+
 				// log::debug("No objects to snap to found");
 			}
 
-			log::debug("refreshing gj controls");
 			GJTransformControl::refreshControl();
-			log::debug("refreshed gj controls, calling updateDisabledWarps");
 			updateDisabledWarps();
 			return std::make_pair(false, nullptr);
 		}