+++ changelog.md
@@ -1,3 +1,7 @@
+# 1.1.1
+- removed the console spam & made the disabling algorithm a bit more reliable
+- thx vitox for reporting a bug im still working on fixing
+
 # 1.1.0
 - Added Apple Macintosh & Adrod Deluxe support.